bot/requirements.txt
python-telegram-bot
python-socketio
python-dotenv
aiohttp
nest_asyncio
httpx
